Free up metrics and uuid fields in a payload if they are still set
These two fields can be allocated by the SparkPlug library without the caller knowing, and so should be freed as well.  This
is a tricky area, though, since it's also possible the caller could have allocated them directly.  We just have to assume
that the caller would have already freed any references and reset the pointers to NULL before calling free_payload if they
had allocated them, and so any references still left in place were the fault of the Sparkplug library and should be cleaned up...