Fix memory leak in usage of grow_char_array
This right here is why people complain about pointers in c.  The main problem that caught my attention was that the way
grow_char_array was being called couldn't possibly update the value of propertyset->keys since it was just passing in the
current pointer value, and not a reference to keys so the function could update it.

Fixing that lead down the path of realizing that the whole body of grow_char_array was doing exactly the right thing to grow
a *sting*, but not an *array of string pointers*.

So, I fixed the whole thing, which required that stupid "char ***" parameter, and renamed the function "grow_string_array" to
be more clear about what was going on.